Move `write_activate_deactivate` implementation and in *main_update.py*
