+++ conda_env/cli/main_update.py
@@ -8,6 +8,7 @@ from conda.cli import common
 from conda.cli import install as cli_install
 from conda.misc import touch_nonadmin
 
+from ..etc_conda_scripts import write_activate_deactivate
 from ..installers.base import get_installer, InvalidInstaller
 from .. import specs as install_specs
 from .. import exceptions
@@ -114,6 +115,8 @@ def execute(args, parser):
                 """).lstrip().format(installer_type)
             )
             return -1
+
+    write_activate_deactivate(env, prefix)
 
     touch_nonadmin(prefix)
     if not args.json:
